Version of Homepage
A redesigned homepage that consists of an about us section, a subscribe to email receiving function for updates on our courses. Also added an animate.css file found online with it's instructions to help with animating the webpage easily and also added a mobile.css to help the website be responsive with smaller mobile screens.